requirements.txt
# QuantByQlib 依赖清单
# Python 3.10.x

# ── UI 框架 ────────────────────────────────────────────────────
PyQt6>=6.7.0
pyqtgraph>=0.13.0

# ── Qlib 生态 ──────────────────────────────────────────────────
pyqlib>=0.9.6

# ── OpenBB Platform（数据统一接入层）──────────────────────────
openbb>=4.0.0
openbb-yfinance>=1.0.0
openbb-alpha-vantage>=1.0.0
openbb-fmp>=1.0.0
openbb-finnhub>=1.0.0
openbb-cboe>=1.0.0

# ── 情绪分析 ──────────────────────────────────────────────────
nltk>=3.8.0
transformers>=4.40.0

# ── 深度学习 ──────────────────────────────────────────────────
torch>=2.3.0
lightgbm>=4.3.0

# ── RD-Agent Docker 集成 ──────────────────────────────────────
docker>=7.0.0

# ── 数据处理 ──────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0

# ── 工具库 ────────────────────────────────────────────────────
pyyaml>=6.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
